server/requirements.txt
pyautogen==0.6.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
requests>=2.32.0
pytest>=8.0.0
black>=24.0.0
isort>=5.13.0
ruff>=0.3.0
pre-commit>=3.7.0
asyncio
dotenv
flaml[automl]
autogen-agentchat
autogen-core
autogen-ext
asyncio
dotenv
openai
tiktoken
autogenstudio
streamlit
arxiv
pillow